requirements.txt
gradio
requests
youtube_transcript_api 
pytube
python-dotenv